Float pricing

+++ api/structs.go
@@ -48,10 +48,10 @@ type Product struct {
 	MiniDescription string    `json:"mini_description"` //Optional
 	Images          []string  `json:"images"`           //Required
 	Available       bool      `json:"available"`        //Required Facetable
-	Price           int       `json:"price"`            //Optional Facetable range | Sortable
+	Price           float64   `json:"price"`      		//Optional Facetable range | Sortable
 	ShipPrice       int       `json:"ship_price"`       //Optional, by default 0
 	Discount        int       `json:"discount"`         //Optional, by default 0
-	FinalPrice      int       `json:"final_price"`      //Optional
+	FinalPrice      float64   `json:"final_price"`      //Optional
 	RelatedProducts []string  `json:"related_products"` //Optional, by default null
 	Brand           string    `json:"brand"`            //Optional, by default "" Facetable
 	Gender          string    `json:"gender"`           //Optional, by default "" Facetable
